+++ awslimitchecker/services/elb.py
@@ -142,16 +142,26 @@ class _ElbService(_AwsService):
         logger.debug(
             'Checking usage for each of %d ALBs', len(lbs['LoadBalancers'])
         )
+        alb_count = 0
+        nlb_count = 0
         for lb in lbs['LoadBalancers']:
-            self._update_usage_for_elbv2(
-                conn2,
-                lb['LoadBalancerArn'],
-                lb['LoadBalancerName']
-            )
+            if lb.get('Type') == 'network':
+                nlb_count += 1
+            else:
+                alb_count += 1
+                self._update_usage_for_alb(
+                    conn2,
+                    lb['LoadBalancerArn'],
+                    lb['LoadBalancerName']
+                )
+        self.limits['Network load balancers']._add_current_usage(
+            nlb_count,
+            aws_type='AWS::ElasticLoadBalancing::NetworkLoadBalancer'
+        )
         logger.debug('Done with ELBv2 usage')
-        return len(lbs['LoadBalancers'])
+        return alb_count
 
-    def _update_usage_for_elbv2(self, conn, alb_arn, alb_name):
+    def _update_usage_for_alb(self, conn, alb_arn, alb_name):
         """
         Update usage for a single ALB.
 
@@ -190,6 +200,32 @@ class _ElbService(_AwsService):
             num_rules,
             aws_type='AWS::ElasticLoadBalancingV2::LoadBalancer',
             resource_id=alb_name,
+        )
+
+    def _update_usage_for_nlb(self, conn, nlb_arn, nlb_name):
+        """
+        Update usage for a single NLB.
+
+        :param conn: elbv2 API connection
+        :type conn: :py:class:`ElasticLoadBalancing.Client`
+        :param nlb_arn: Load Balancer ARN
+        :type nlb_arn: str
+        :param nlb_name: Load Balancer Name
+        :type nlb_name: str
+        """
+        logger.debug('Updating usage for NLB %s', nlb_arn)
+        listeners = paginate_dict(
+            conn.describe_listeners,
+            LoadBalancerArn=nlb_arn,
+            alc_marker_path=['NextMarker'],
+            alc_data_path=['Listeners'],
+            alc_marker_param='Marker'
+        )['Listeners']
+        self.limits[
+            'Listeners per network load balancer']._add_current_usage(
+            len(listeners),
+            aws_type='AWS::ElasticLoadBalancingV2::NetworkLoadBalancer',
+            resource_id=nlb_name
         )
 
     def get_limits(self):
@@ -249,6 +285,23 @@ class _ElbService(_AwsService):
             limit_type='AWS::ElasticLoadBalancingV2::LoadBalancer',
             limit_subtype='LoadBalancerRule'
         )
+        limits['Network load balancers'] = AwsLimit(
+            'Network load balancers',
+            self,
+            20,
+            self.warning_threshold,
+            self.critical_threshold,
+            limit_type='AWS::ElasticLoadBalancing::NetworkLoadBalancer',
+        )
+        limits['Listeners per network load balancer'] = AwsLimit(
+            'Listeners per network load balancer',
+            self,
+            50,
+            self.warning_threshold,
+            self.critical_threshold,
+            limit_type='AWS::ElasticLoadBalancingV2::NetworkLoadBalancer',
+            limit_subtype='LoadBalancerListener'
+        )
         self.limits = limits
         return limits
 
@@ -282,7 +335,10 @@ class _ElbService(_AwsService):
             'listeners-per-application-load-balancer':
                 'Listeners per application load balancer',
             'rules-per-application-load-balancer':
-                'Rules per application load balancer'
+                'Rules per application load balancer',
+            'network-load-balancers': 'Network load balancers',
+            'listeners-per-network-load-balancer':
+                'Listeners per network load balancer'
         }
         for attrib in attribs['Limits']:
             if int(attrib.get('Max', 0)) == 0: