+++ awslimitchecker/tests/services/test_elb.py
@@ -74,6 +74,8 @@ class Test_ElbService(object):
             'Active load balancers',
             'Listeners per application load balancer',
             'Listeners per load balancer',
+            'Listeners per network load balancer',
+            'Network load balancers',
             'Rules per application load balancer',
             'Target groups'
         ])
@@ -121,6 +123,9 @@ class Test_ElbService(object):
         assert cls.limits[
             'Listeners per application load balancer'].api_limit == 9
         assert cls.limits['Rules per application load balancer'].api_limit == 10
+        assert cls.limits[
+            'Listeners per network load balancer'].api_limit == 100
+        assert cls.limits['Network load balancers'].api_limit == 40
 
     def test_find_usage(self):
         with patch('%s._find_usage_elbv1' % pb, autospec=True) as mock_v1:
@@ -181,7 +186,7 @@ class Test_ElbService(object):
                     PropertyMock(return_value='rname')
                 with patch('%s.paginate_dict' % pbm) as mock_paginate:
                     with patch(
-                        '%s._update_usage_for_elbv2' % pb, autospec=True
+                        '%s._update_usage_for_alb' % pb, autospec=True
                     ) as mock_u:
                         with patch(
                             '%s.Config' % pbm, autospec=True
@@ -226,18 +231,23 @@ class Test_ElbService(object):
         assert len(lim) == 1
         assert lim[0].get_value() == 3
         assert lim[0].aws_type == 'AWS::ElasticLoadBalancingV2::TargetGroup'
+        lim = cls.limits['Network load balancers'].get_current_usage()
+        assert len(lim) == 1
+        assert lim[0].get_value() == 1
+        assert lim[0].aws_type == \
+            'AWS::ElasticLoadBalancing::NetworkLoadBalancer'
 
-    def test_update_usage_for_elbv2(self):
+    def test_update_usage_for_alb(self):
         conn = Mock()
         with patch('%s.paginate_dict' % pbm) as mock_paginate:
             mock_paginate.side_effect = [
-                result_fixtures.ELB.test_usage_elbv2_listeners,
-                result_fixtures.ELB.test_usage_elbv2_rules[0],
-                result_fixtures.ELB.test_usage_elbv2_rules[1],
-                result_fixtures.ELB.test_usage_elbv2_rules[2]
+                result_fixtures.ELB.test_usage_alb_listeners,
+                result_fixtures.ELB.test_usage_alb_rules[0],
+                result_fixtures.ELB.test_usage_alb_rules[1],
+                result_fixtures.ELB.test_usage_alb_rules[2]
             ]
             cls = _ElbService(21, 43)
-            cls._update_usage_for_elbv2(conn, 'myarn', 'albname')
+            cls._update_usage_for_alb(conn, 'myarn', 'albname')
         assert mock_paginate.mock_calls == [
             call(
                 conn.describe_listeners,
@@ -268,18 +278,43 @@ class Test_ElbService(object):
                 alc_marker_param='Marker'
             )
         ]
-        l = cls.limits[
+        lim = cls.limits[
             'Listeners per application load balancer'].get_current_usage()
-        assert len(l) == 1
-        assert l[0].get_value() == 3
-        assert l[0].aws_type == 'AWS::ElasticLoadBalancingV2::LoadBalancer'
-        assert l[0].resource_id == 'albname'
+        assert len(lim) == 1
+        assert lim[0].get_value() == 3
+        assert lim[0].aws_type == 'AWS::ElasticLoadBalancingV2::LoadBalancer'
+        assert lim[0].resource_id == 'albname'
         r = cls.limits[
             'Rules per application load balancer'].get_current_usage()
         assert len(r) == 1
         assert r[0].get_value() == 7
         assert r[0].aws_type == 'AWS::ElasticLoadBalancingV2::LoadBalancer'
         assert r[0].resource_id == 'albname'
+
+    def test_update_usage_for_nlb(self):
+        conn = Mock()
+        with patch('%s.paginate_dict' % pbm) as mock_paginate:
+            mock_paginate.side_effect = [
+                result_fixtures.ELB.test_usage_nlb_listeners
+            ]
+            cls = _ElbService(21, 43)
+            cls._update_usage_for_nlb(conn, 'mynarn', 'nlbname')
+        assert mock_paginate.mock_calls == [
+            call(
+                conn.describe_listeners,
+                LoadBalancerArn='mynarn',
+                alc_marker_path=['NextMarker'],
+                alc_data_path=['Listeners'],
+                alc_marker_param='Marker'
+            )
+        ]
+        lim = cls.limits[
+            'Listeners per network load balancer'].get_current_usage()
+        assert len(lim) == 1
+        assert lim[0].get_value() == 2
+        assert lim[0].aws_type == \
+            'AWS::ElasticLoadBalancingV2::NetworkLoadBalancer'
+        assert lim[0].resource_id == 'nlbname'
 
     def test_required_iam_permissions(self):
         cls = _ElbService(21, 43)